requirements.txt
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-readers-file
openai>=1.40.0
python-dotenv==1.0.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
pandas>=2.0.0
numpy>=1.24.0
